Update batch_ingest_sm_pyspark.py
Null feature values can't be ingested into a feature group. This line of code stops the code from following error:

Attempted to parse the feature value for the feature named [z] into a FeatureValue of type Fractional. The provided value must be within the range of a double precision floating point number defined by the IEEE 754 standard. The input format can be in either decimal form or scientific notation.